asserts: start splitting out attrMatcher for reuse to constraint.go
