+++ asserts/ifacedecls_test.go
@@ -1,7 +1,7 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 
 /*
- * Copyright (C) 2015-2017 Canonical Ltd
+ * Copyright (C) 2015-2022 Canonical Ltd
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -118,43 +118,14 @@ func (s *attrConstraintsSuite) TestSimple(c *C) {
 	c.Check(err, ErrorMatches, `attribute "bar" has constraints but is unset`)
 }
 
-func (s *attrConstraintsSuite) TestSimpleAnchorsVsRegexpAlt(c *C) {
+func (s *attrConstraintsSuite) TestMissingCheck(c *C) {
 	m, err := asserts.ParseHeaders([]byte(`attrs:
-  bar: BAR|BAZ`))
+  foo: $MISSING`))
 	c.Assert(err, IsNil)
 
 	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
 	c.Assert(err, IsNil)
-
-	plug := attrerObject(map[string]interface{}{
-		"bar": "BAR",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, IsNil)
-
-	plug = attrerObject(map[string]interface{}{
-		"bar": "BARR",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, ErrorMatches, `attribute "bar" value "BARR" does not match \^\(BAR|BAZ\)\$`)
-
-	plug = attrerObject(map[string]interface{}{
-		"bar": "BBAZ",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, ErrorMatches, `attribute "bar" value "BAZZ" does not match \^\(BAR|BAZ\)\$`)
-
-	plug = attrerObject(map[string]interface{}{
-		"bar": "BABAZ",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, ErrorMatches, `attribute "bar" value "BABAZ" does not match \^\(BAR|BAZ\)\$`)
-
-	plug = attrerObject(map[string]interface{}{
-		"bar": "BARAZ",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, ErrorMatches, `attribute "bar" value "BARAZ" does not match \^\(BAR|BAZ\)\$`)
+	c.Check(asserts.RuleFeature(cstrs, "dollar-attr-constraints"), Equals, true)
 }
 
 func (s *attrConstraintsSuite) TestNested(c *C) {
@@ -205,111 +176,6 @@ bar:
 baz: BAZ
 `), nil)
 	c.Check(err, ErrorMatches, `attribute "bar\.bar2" must be a scalar or list`)
-}
-
-func (s *attrConstraintsSuite) TestAlternative(c *C) {
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  -
-    foo: FOO
-    bar: BAR
-  -
-    foo: FOO
-    bar: BAZ`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].([]interface{}))
-	c.Assert(err, IsNil)
-
-	plug := attrerObject(map[string]interface{}{
-		"foo": "FOO",
-		"bar": "BAR",
-		"baz": "BAZ",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, IsNil)
-
-	plug = attrerObject(map[string]interface{}{
-		"foo": "FOO",
-		"bar": "BAZ",
-		"baz": "BAZ",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, IsNil)
-
-	plug = attrerObject(map[string]interface{}{
-		"foo": "FOO",
-		"bar": "BARR",
-		"baz": "BAR",
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, ErrorMatches, `no alternative matches: attribute "bar" value "BARR" does not match \^\(BAR\)\$`)
-}
-
-func (s *attrConstraintsSuite) TestNestedAlternative(c *C) {
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  foo: FOO
-  bar:
-    bar1: BAR1
-    bar2:
-      - BAR2
-      - BAR22`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: FOO
-bar:
-  bar1: BAR1
-  bar2: BAR2
-`), nil)
-	c.Check(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: FOO
-bar:
-  bar1: BAR1
-  bar2: BAR22
-`), nil)
-	c.Check(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: FOO
-bar:
-  bar1: BAR1
-  bar2: BAR3
-`), nil)
-	c.Check(err, ErrorMatches, `no alternative for attribute "bar\.bar2" matches: attribute "bar\.bar2" value "BAR3" does not match \^\(BAR2\)\$`)
-}
-
-func (s *attrConstraintsSuite) TestAlternativeMatchingStringList(c *C) {
-	toMatch := attrs(`
-write:
- - /var/tmp
- - /var/lib/snapd/snapshots
-`)
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  write: /var/(tmp|lib/snapd/snapshots)`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-
-	err = cstrs.Check(toMatch, nil)
-	c.Check(err, IsNil)
-
-	m, err = asserts.ParseHeaders([]byte(`attrs:
-  write:
-    - /var/tmp
-    - /var/lib/snapd/snapshots`))
-	c.Assert(err, IsNil)
-
-	cstrsLst, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-
-	err = cstrsLst.Check(toMatch, nil)
-	c.Check(err, IsNil)
 }
 
 func (s *attrConstraintsSuite) TestAlternativeMatchingComplex(c *C) {
@@ -390,13 +256,6 @@ foo: 1
 bar: true
 `), nil)
 	c.Check(err, IsNil)
-
-	plug := attrerObject(map[string]interface{}{
-		"foo": int64(1),
-		"bar": true,
-	})
-	err = cstrs.Check(plug, nil)
-	c.Check(err, IsNil)
 }
 
 func (s *attrConstraintsSuite) TestCompileErrors(c *C) {
@@ -404,16 +263,6 @@ func (s *attrConstraintsSuite) TestCompileErrors(c *C) {
 		"foo": "[",
 	})
 	c.Check(err, ErrorMatches, `cannot compile "foo" constraint "\[": error parsing regexp:.*`)
-
-	_, err = asserts.CompileAttributeConstraints(map[string]interface{}{
-		"foo": []interface{}{"foo", "["},
-	})
-	c.Check(err, ErrorMatches, `cannot compile "foo/alt#2/" constraint "\[": error parsing regexp:.*`)
-
-	_, err = asserts.CompileAttributeConstraints(map[string]interface{}{
-		"foo": []interface{}{"foo", []interface{}{"bar", "baz"}},
-	})
-	c.Check(err, ErrorMatches, `cannot nest alternative constraints directly at "foo/alt#2/"`)
 
 	_, err = asserts.CompileAttributeConstraints("FOO")
 	c.Check(err, ErrorMatches, `first level of non alternative constraints must be a set of key-value contraints`)
@@ -435,45 +284,6 @@ func (s *attrConstraintsSuite) TestCompileErrors(c *C) {
 		c.Check(err, ErrorMatches, fmt.Sprintf(`cannot compile "foo" constraint "%s": not a valid \$SLOT\(\)/\$PLUG\(\) constraint`, regexp.QuoteMeta(wrong)))
 
 	}
-}
-
-func (s *attrConstraintsSuite) TestMatchingListsSimple(c *C) {
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  foo: /foo/.*`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: ["/foo/x", "/foo/y"]
-`), nil)
-	c.Check(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: ["/foo/x", "/foo"]
-`), nil)
-	c.Check(err, ErrorMatches, `attribute "foo\.1" value "/foo" does not match \^\(/foo/\.\*\)\$`)
-}
-
-func (s *attrConstraintsSuite) TestMissingCheck(c *C) {
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  foo: $MISSING`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-	c.Check(asserts.RuleFeature(cstrs, "dollar-attr-constraints"), Equals, true)
-
-	err = cstrs.Check(attrs(`
-bar: baz
-`), nil)
-	c.Check(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: ["x"]
-`), nil)
-	c.Check(err, ErrorMatches, `attribute "foo" is constrained to be missing but is set`)
 }
 
 type testEvalAttr struct {
@@ -546,30 +356,6 @@ foo: foo
 bar: bar.baz
 `), testEvalAttr{comp3})
 	c.Check(err, ErrorMatches, `attribute "foo" does not match \$SLOT\(foo\): foo != other-value`)
-}
-
-func (s *attrConstraintsSuite) TestMatchingListsMap(c *C) {
-	m, err := asserts.ParseHeaders([]byte(`attrs:
-  foo:
-    p: /foo/.*`))
-	c.Assert(err, IsNil)
-
-	cstrs, err := asserts.CompileAttributeConstraints(m["attrs"].(map[string]interface{}))
-	c.Assert(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: [{p: "/foo/x"}, {p: "/foo/y"}]
-`), nil)
-	c.Check(err, IsNil)
-
-	err = cstrs.Check(attrs(`
-foo: [{p: "zzz"}, {p: "/foo/y"}]
-`), nil)
-	c.Check(err, ErrorMatches, `attribute "foo\.0\.p" value "zzz" does not match \^\(/foo/\.\*\)\$`)
-}
-
-func (s *attrConstraintsSuite) TestAlwaysMatchAttributeConstraints(c *C) {
-	c.Check(asserts.AlwaysMatchAttributes.Check(nil, nil), IsNil)
 }
 
 func (s *attrConstraintsSuite) TestNeverMatchAttributeConstraints(c *C) {
